Major update on scripts, folders and menu options

+++ mainconftree/installzsh.sh
@@ -5,8 +5,7 @@ clear
 		echo "$(tput bold)$(tput setaf 7)"
         echo "|==============|> MAINCONF <|==============|"
         echo "|--------------> INSTALL ZSH <-------------|"
-		sleep 1
-
+		
 		read -p "|--> Insert your username: " username	
 
 		if [ -z "$(pacman -Qi zsh)" ]; then sudo pacman -S zsh --needed --noconfirm; else echo "|--> zsh is installed" && sleep 1; fi
@@ -20,29 +19,15 @@ clear
 
 	    #installing zsh-autosuggestions plugin
 		echo "|--> Installing auto suggestions plugin" ; sleep 1
-		if [ -d ~/.oh-my-zsh/custom/plugins/zsh-autosuggestions ]; then
-		sudo rm -rf ~/.oh-my-zsh/custom/plugins/zsh-autosuggestions
-		fi
+		if [ -d ~/.oh-my-zsh/custom/plugins/zsh-autosuggestions ]; then sudo rm -rf ~/.oh-my-zsh/custom/plugins/zsh-autosuggestions; fi
 		git clone --quiet https://github.com/zsh-users/zsh-autosuggestions ~/.oh-my-zsh/custom/plugins/zsh-autosuggestions
-		echo "|--> ZSH Autosuggestions installed!!" ; sleep 1
 		
 		#installing powerline fonts
 		echo "|--> Downloading Powerline fonts" ; sleep 1
 		git clone --quiet https://github.com/powerline/fonts.git $HOME/.local/share/fonts/
 		$HOME/.local/share/fonts/install.sh
-		echo "|--> Powerline fonts installed!!" ; sleep 1
 	
 		#installing pi theme https://github.com/tobyjamesthomas/pi
 		echo "|--> Installing zsh pi theme" ; sleep 1
-	    if [ -d ~/.oh-my-zsh/custom/themes/pi ]; then
-		sudo rm -rf ~/.oh-my-zsh/custom/themes/pi
-		fi
-	    
+	    if [ -d ~/.oh-my-zsh/custom/themes/pi ]; then sudo rm -rf ~/.oh-my-zsh/custom/themes/pi; fi
 		wget -O ~/.oh-my-zsh/custom/themes/pi.zsh-theme https://raw.githubusercontent.com/joaov777/mainconf/master/zsh/pi.zsh-theme
-		echo "|--> ZSH pi theme installed!!" ; sleep 1
-
-		
-	
-		
-	    
-
